feat(preview): add use-origin hook to get the origin of the current window location
This hook first checks if the component is mounted on the client-side where the window object is available. If it's not mounted (likely server-side rendering), it returns an empty string. If it is mounted (client-side), it returns the window.location.origin. This ensures compatibility with both client-side and server-side rendering